Handled exceptions

+++ backend/bank/src/main/java/com/hasib/bank/model/Transaction.java
@@ -17,14 +17,17 @@ public class Transaction {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
+
     @JsonManagedReference
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "sender_id")
     private UserEntity sender;
+
     @JsonManagedReference
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "receiver_id")
     private UserEntity receiver;
-    private long amount;
+
+    private double amount;
     private LocalDateTime createdAt;
 }
